AvengersBattleSimulation
Avengers are divided into heroes and enemies. Quantum Nano Suits are alloted to them. State of avenger and suit are updated after every attack. Suits can be upgraded, boosted, repaired. Battle log notes all the attacks that took place in the battle. The team with more power and durability at the end of battle is the wining team.