CourseRegistrationSystem
Course Registration System allots courses to students based on their eligibility. Students can add/drop courses. The instructor is warned if there is a cyclic dependency in the prerequisites of a course. If the course capacity is met, waitlist is maintained, if some student drops the course, it is allotted to student in the waitlist on first come first serve basis.